+++ templates/slots/Services/COPage/PageComponent/classes/class.il__PLUGIN_NAME__Plugin.php
@@ -93,7 +93,7 @@ class il__PLUGIN_NAME__Plugin extends ilPageComponentPlugin
      */
     protected function shouldUseOneUpdateStepOnly() : bool
     {
-        return true;
+        return __SHOULD_USE_ONE_UPDATE_STEP_ONLY__;
     }